warn users about the broken password store
waybackwhen, we decided that it's an advantage to give the users
their original password if they have it forgotten.

Of course, this is totally broken - especially considering the
skill-level of the programmer who has initially written this stuff.

For now, let's just warn the user. I'll have to fix this for sure
later on